requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.3.0=pypi_0
anyio=3.5.0=py310haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py310haa95532_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2022.10.11=haa95532_0
cachetools=5.2.0=pypi_0
certifi=2022.9.24=py310haa95532_0
cffi=1.15.1=py310h2bbff1b_2
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
colorama=0.4.5=py310haa95532_0
contourpy=1.0.6=pypi_0
cryptography=38.0.1=py310h21b164f_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flask=2.2.2=pypi_0
flask-cors=3.0.10=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=ha860e81_0
gitdb=4.0.10=pypi_0
gitpython=3.1.29=pypi_0
glib=2.69.1=h5dc1a3c_2
google-auth=2.14.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.50.0=pypi_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.2=py310haa95532_0
ipython=8.6.0=py310haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
itsdangerous=2.1.2=pypi_0
jedi=0.18.1=py310haa95532_1
jinja2=3.1.2=py310haa95532_0
jpeg=9e=h2bbff1b_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.16.0=py310haa95532_0
jupyter=1.0.0=py310haa95532_8
jupyter_client=7.4.7=py310haa95532_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.11.2=py310haa95532_0
jupyter_server=1.18.1=py310haa95532_0
jupyterlab=3.5.0=py310haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.16.3=py310haa95532_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=pypi_0
lerc=3.0=hd77b12b_0
libclang=12.0.0=default_h627e005_2
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.4.0=h8a3f274_2
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lxml=4.9.1=py310h1985fb9_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=py310h2bbff1b_0
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=py310haa95532_0
mistune=0.8.4=py310h2bbff1b_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
nbclassic=0.4.8=py310haa95532_0
nbclient=0.5.13=py310haa95532_0
nbconvert=6.5.4=py310haa95532_0
nbformat=5.5.0=py310haa95532_0
nest-asyncio=1.5.5=py310haa95532_0
notebook=6.5.2=py310haa95532_0
notebook-shim=0.2.2=py310haa95532_0
nsight-compute=2022.3.0.22=0
numpy=1.23.5=pypi_0
numpy-base=1.23.4=py310h04254f7_0
oauthlib=3.2.2=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1s=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.2=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.3.0=pypi_0
pip=22.2.2=py310haa95532_0
ply=3.11=py310haa95532_0
prometheus_client=0.14.1=py310haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.20.3=pypi_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310haa95532_0
pyqt=5.15.7=py310hd77b12b_0
pyqt5-sip=12.11.0=py310hd77b12b_0
pyrsistent=0.18.0=py310h2bbff1b_0
pysocks=1.7.1=py310haa95532_0
python=3.10.8=h966fe2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py310haa95532_0
pytorch=1.13.0=py3.10_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h67b0de4_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py310haa95532_0
pywin32=302=py310h2bbff1b_2
pywinpty=2.0.2=py310h5da7b33_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py310hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtconsole=5.3.2=py310haa95532_0
qtpy=2.2.0=py310haa95532_0
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=py310haa95532_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.9.3=pypi_0
seaborn=0.12.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.5.0=py310haa95532_0
sip=6.6.2=py310hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.2.0=py310haa95532_1
soupsieve=2.3.2.post1=py310haa95532_0
sqlite=3.40.0=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.13.1=py310haa95532_0
thop=0.1.1-2209072238=pypi_0
tinycss2=1.2.1=py310haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py310haa95532_0
torch=1.13.0=pypi_0
torchaudio=0.13.0=pypi_0
torchvision=0.14.0=pypi_0
tornado=6.2=py310h2bbff1b_0
tqdm=4.64.1=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py310haa95532_0
typing_extensions=4.3.0=py310haa95532_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.12=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310haa95532_1
websocket-client=0.58.0=py310haa95532_4
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
wincertstore=0.2=py310haa95532_2
winpty=0.4.3=4
xz=5.2.6=h8cc25b3_0
zeromq=4.3.4=hd77b12b_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
